Refactor: separate blockchain from lib.rs
